Refactor Reconfigure to break down into smaller parts
